add screen more
display home screen
