+++ .expo/devices.json
@@ -1,0 +1,3 @@
+{
+  "devices": []
+}
