+++ src/screens/index.js
@@ -1,0 +1,6 @@
+import Home from "./Home";
+import Launch from "./Launch";
+import Profile from "./Profile";
+import Statistics from "./Statistics";
+
+export {Home, Launch, Profile, Statistics};
